+++ private/sourceRange.mjs
@@ -6,7 +6,7 @@
  * @param {string} source Source code.
  * @param {number} startOffset Start character offset.
  * @param {number} endOffset End character offset.
- * @param {string|boolean} [ignoreNextLineComment=' coverage ignore next line'] Single line case-insensitive comment content to ignore ranges that start on the the next line, or `false` to disable ignore comments.
+ * @param {string|boolean} [ignoreNextLineComment=" coverage ignore next line"] Single line case-insensitive comment content to ignore ranges that start on the the next line, or `false` to disable ignore comments.
  * @returns {SourceCodeRange} Source code range info.
  * @ignore
  */
@@ -14,23 +14,23 @@ export default function sourceRange(
   source,
   startOffset,
   endOffset,
-  ignoreNextLineComment = ' coverage ignore next line'
+  ignoreNextLineComment = " coverage ignore next line"
 ) {
-  if (typeof source !== 'string')
-    throw new TypeError('First argument `source` must be a string.');
+  if (typeof source !== "string")
+    throw new TypeError("First argument `source` must be a string.");
 
-  if (typeof startOffset !== 'number')
-    throw new TypeError('Second argument `startOffset` must be a number.');
+  if (typeof startOffset !== "number")
+    throw new TypeError("Second argument `startOffset` must be a number.");
 
-  if (typeof endOffset !== 'number')
-    throw new TypeError('Third argument `endOffset` must be a number.');
+  if (typeof endOffset !== "number")
+    throw new TypeError("Third argument `endOffset` must be a number.");
 
   if (
-    typeof ignoreNextLineComment !== 'string' &&
+    typeof ignoreNextLineComment !== "string" &&
     ignoreNextLineComment !== false
   )
     throw new TypeError(
-      'Fourth argument `ignoreNextLineComment` must be a string or `false`.'
+      "Fourth argument `ignoreNextLineComment` must be a string or `false`."
     );
 
   if (ignoreNextLineComment)
@@ -48,7 +48,7 @@ export default function sourceRange(
     const nextLineOffset = lineOffset + lineSource.length;
 
     if (
-      !('line' in range.start) &&
+      !("line" in range.start) &&
       startOffset >= lineOffset &&
       startOffset < nextLineOffset
     ) {