+++ test/cli/coverage-node.test.mjs
@@ -1,28 +1,28 @@
-import { strictEqual } from 'assert';
-import { spawnSync } from 'child_process';
-import fs from 'fs';
-import { join, relative } from 'path';
-import { fileURLToPath } from 'url';
-import disposableDirectory from 'disposable-directory';
-import snapshot from 'snapshot-assertion';
-import coverageSupported from '../../public/coverageSupported.mjs';
-import replaceStackTraces from '../replaceStackTraces.mjs';
+import { strictEqual } from "assert";
+import { spawnSync } from "child_process";
+import fs from "fs";
+import { join, relative } from "path";
+import { fileURLToPath } from "url";
+import disposableDirectory from "disposable-directory";
+import snapshot from "snapshot-assertion";
+import coverageSupported from "../../public/coverageSupported.mjs";
+import replaceStackTraces from "../replaceStackTraces.mjs";
 
 const COVERAGE_NODE_CLI_PATH = fileURLToPath(
-  new URL('../../cli/coverage-node.mjs', import.meta.url)
+  new URL("../../cli/coverage-node.mjs", import.meta.url)
 );
-const SNAPSHOT_REPLACEMENT_FILE_PATH = '<file path>';
-const SNAPSHOT_REPLACEMENT_PROCESS_NODE_VERSION = '<process Node.js version>';
+const SNAPSHOT_REPLACEMENT_FILE_PATH = "<file path>";
+const SNAPSHOT_REPLACEMENT_PROCESS_NODE_VERSION = "<process Node.js version>";
 
 export default (tests) => {
-  tests.add('`coverage-node` CLI with 1 covered file.', async () => {
+  tests.add("`coverage-node` CLI with 1 covered file.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const filePath = join(tempDirPath, 'index.mjs');
+      const filePath = join(tempDirPath, "index.mjs");
 
-      await fs.promises.writeFile(filePath, '1;');
+      await fs.promises.writeFile(filePath, "1;");
 
       const { stdout, stderr, status, error } = spawnSync(
-        'node',
+        "node",
         [COVERAGE_NODE_CLI_PATH, filePath],
         {
           env: {
@@ -38,7 +38,7 @@ export default (tests) => {
         coverageSupported
           ? stdout
               .toString()
-              .replace(relative('', filePath), SNAPSHOT_REPLACEMENT_FILE_PATH)
+              .replace(relative("", filePath), SNAPSHOT_REPLACEMENT_FILE_PATH)
           : stdout
               .toString()
               .replace(
@@ -47,20 +47,20 @@ export default (tests) => {
               ),
         new URL(
           `../snapshots/coverage-node/1-covered-file-coverage-${
-            coverageSupported ? 'supported' : 'unsupported'
+            coverageSupported ? "supported" : "unsupported"
           }-stdout.ans`,
           import.meta.url
         )
       );
 
-      strictEqual(stderr.toString(), '');
+      strictEqual(stderr.toString(), "");
       strictEqual(status, 0);
     });
   });
 
-  tests.add('`coverage-node` CLI with 1 ignored file.', async () => {
+  tests.add("`coverage-node` CLI with 1 ignored file.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const filePath = join(tempDirPath, 'index.mjs');
+      const filePath = join(tempDirPath, "index.mjs");
 
       await fs.promises.writeFile(
         filePath,
@@ -70,7 +70,7 @@ export default (tests) => {
       );
 
       const { stdout, stderr, status, error } = spawnSync(
-        'node',
+        "node",
         [COVERAGE_NODE_CLI_PATH, filePath],
         {
           env: {
@@ -86,7 +86,7 @@ export default (tests) => {
         coverageSupported
           ? stdout
               .toString()
-              .replace(relative('', filePath), SNAPSHOT_REPLACEMENT_FILE_PATH)
+              .replace(relative("", filePath), SNAPSHOT_REPLACEMENT_FILE_PATH)
           : stdout
               .toString()
               .replace(
@@ -95,25 +95,25 @@ export default (tests) => {
               ),
         new URL(
           `../snapshots/coverage-node/1-ignored-file-coverage-${
-            coverageSupported ? 'supported' : 'unsupported'
+            coverageSupported ? "supported" : "unsupported"
           }-stdout.ans`,
           import.meta.url
         )
       );
 
-      strictEqual(stderr.toString(), '');
+      strictEqual(stderr.toString(), "");
       strictEqual(status, 0);
     });
   });
 
-  tests.add('`coverage-node` CLI with 1 uncovered file.', async () => {
+  tests.add("`coverage-node` CLI with 1 uncovered file.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const filePath = join(tempDirPath, 'index.mjs');
+      const filePath = join(tempDirPath, "index.mjs");
 
-      await fs.promises.writeFile(filePath, '() => {};');
+      await fs.promises.writeFile(filePath, "() => {};");
 
       const { stdout, stderr, status, error } = spawnSync(
-        'node',
+        "node",
         [COVERAGE_NODE_CLI_PATH, filePath],
         {
           env: {
@@ -136,7 +136,7 @@ export default (tests) => {
               ),
         new URL(
           `../snapshots/coverage-node/1-uncovered-file-coverage-${
-            coverageSupported ? 'supported' : 'unsupported'
+            coverageSupported ? "supported" : "unsupported"
           }-stdout.ans`,
           import.meta.url
         )
@@ -146,39 +146,39 @@ export default (tests) => {
         await snapshot(
           stderr
             .toString()
-            .replace(relative('', filePath), SNAPSHOT_REPLACEMENT_FILE_PATH),
+            .replace(relative("", filePath), SNAPSHOT_REPLACEMENT_FILE_PATH),
           new URL(
-            '../snapshots/coverage-node/1-uncovered-file-coverage-supported-stderr.ans',
+            "../snapshots/coverage-node/1-uncovered-file-coverage-supported-stderr.ans",
             import.meta.url
           )
         );
-      else strictEqual(stderr.toString(), '');
+      else strictEqual(stderr.toString(), "");
 
       strictEqual(status, coverageSupported ? 1 : 0);
     });
   });
 
   tests.add(
-    '`coverage-node` CLI with 2 covered, ignored and uncovered files.',
+    "`coverage-node` CLI with 2 covered, ignored and uncovered files.",
     async () => {
       await disposableDirectory(async (tempDirPath) => {
-        const fileAPath = join(tempDirPath, 'a.mjs');
-        const fileBPath = join(tempDirPath, 'b.mjs');
-        const fileCPath = join(tempDirPath, 'c.mjs');
-        const fileDPath = join(tempDirPath, 'd.mjs');
-        const fileEPath = join(tempDirPath, 'e.mjs');
-        const fileFPath = join(tempDirPath, 'f.mjs');
+        const fileAPath = join(tempDirPath, "a.mjs");
+        const fileBPath = join(tempDirPath, "b.mjs");
+        const fileCPath = join(tempDirPath, "c.mjs");
+        const fileDPath = join(tempDirPath, "d.mjs");
+        const fileEPath = join(tempDirPath, "e.mjs");
+        const fileFPath = join(tempDirPath, "f.mjs");
 
         await fs.promises.writeFile(
           fileAPath,
-          `import '${fileBPath}';
-import '${fileCPath}';
-import '${fileDPath}';
-import '${fileEPath}';
-import '${fileFPath}';
+          `import "${fileBPath}";
+import "${fileCPath}";
+import "${fileDPath}";
+import "${fileEPath}";
+import "${fileFPath}";
 `
         );
-        await fs.promises.writeFile(fileBPath, 'function a() {}; a();');
+        await fs.promises.writeFile(fileBPath, "function a() {}; a();");
         await fs.promises.writeFile(
           fileCPath,
           `// coverage ignore next line
@@ -189,11 +189,11 @@ import '${fileFPath}';
           `// coverage ignore next line
 () => {};`
         );
-        await fs.promises.writeFile(fileEPath, '() => {};');
-        await fs.promises.writeFile(fileFPath, '() => {};');
+        await fs.promises.writeFile(fileEPath, "() => {};");
+        await fs.promises.writeFile(fileFPath, "() => {};");
 
         const { stdout, stderr, status, error } = spawnSync(
-          'node',
+          "node",
           [COVERAGE_NODE_CLI_PATH, fileAPath],
           {
             env: {
@@ -209,10 +209,10 @@ import '${fileFPath}';
           coverageSupported
             ? stdout
                 .toString()
-                .replace(relative('', fileAPath), '<pathA>')
-                .replace(relative('', fileBPath), '<pathB>')
-                .replace(relative('', fileCPath), '<pathC>')
-                .replace(relative('', fileDPath), '<pathD>')
+                .replace(relative("", fileAPath), "<pathA>")
+                .replace(relative("", fileBPath), "<pathB>")
+                .replace(relative("", fileCPath), "<pathC>")
+                .replace(relative("", fileDPath), "<pathD>")
             : stdout
                 .toString()
                 .replace(
@@ -221,7 +221,7 @@ import '${fileFPath}';
                 ),
           new URL(
             `../snapshots/coverage-node/2-covered-ignored-uncovered-files-coverage-${
-              coverageSupported ? 'supported' : 'unsupported'
+              coverageSupported ? "supported" : "unsupported"
             }-stdout.ans`,
             import.meta.url
           )
@@ -231,28 +231,28 @@ import '${fileFPath}';
           await snapshot(
             stderr
               .toString()
-              .replace(relative('', fileEPath), '<pathE>')
-              .replace(relative('', fileFPath), '<pathF>'),
+              .replace(relative("", fileEPath), "<pathE>")
+              .replace(relative("", fileFPath), "<pathF>"),
             new URL(
-              '../snapshots/coverage-node/2-covered-ignored-uncovered-files-coverage-supported-stderr.ans',
+              "../snapshots/coverage-node/2-covered-ignored-uncovered-files-coverage-supported-stderr.ans",
               import.meta.url
             )
           );
-        else strictEqual(stderr.toString(), '');
+        else strictEqual(stderr.toString(), "");
 
         strictEqual(status, coverageSupported ? 1 : 0);
       });
     }
   );
 
-  tests.add('`coverage-node` CLI with a script console log.', async () => {
+  tests.add("`coverage-node` CLI with a script console log.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const filePath = join(tempDirPath, 'index.mjs');
+      const filePath = join(tempDirPath, "index.mjs");
 
-      await fs.promises.writeFile(filePath, "console.log('Message.');");
+      await fs.promises.writeFile(filePath, 'console.log("Message.");');
 
       const { stdout, stderr, status, error } = spawnSync(
-        'node',
+        "node",
         [COVERAGE_NODE_CLI_PATH, filePath],
         {
           env: {
@@ -268,7 +268,7 @@ import '${fileFPath}';
         coverageSupported
           ? stdout
               .toString()
-              .replace(relative('', filePath), SNAPSHOT_REPLACEMENT_FILE_PATH)
+              .replace(relative("", filePath), SNAPSHOT_REPLACEMENT_FILE_PATH)
           : stdout
               .toString()
               .replace(
@@ -277,25 +277,25 @@ import '${fileFPath}';
               ),
         new URL(
           `../snapshots/coverage-node/script-console-log-coverage-${
-            coverageSupported ? 'supported' : 'unsupported'
+            coverageSupported ? "supported" : "unsupported"
           }-stdout.ans`,
           import.meta.url
         )
       );
 
-      strictEqual(stderr.toString(), '');
+      strictEqual(stderr.toString(), "");
       strictEqual(status, 0);
     });
   });
 
-  tests.add('`coverage-node` CLI with a script error.', async () => {
+  tests.add("`coverage-node` CLI with a script error.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const filePath = join(tempDirPath, 'index.mjs');
+      const filePath = join(tempDirPath, "index.mjs");
 
-      await fs.promises.writeFile(filePath, "throw new Error('Error.');");
+      await fs.promises.writeFile(filePath, 'throw new Error("Error.");');
 
       const { stdout, stderr, status, error } = spawnSync(
-        'node',
+        "node",
         [COVERAGE_NODE_CLI_PATH, filePath],
         {
           env: {
@@ -307,7 +307,7 @@ import '${fileFPath}';
 
       if (error) throw error;
 
-      strictEqual(stdout.toString(), '');
+      strictEqual(stdout.toString(), "");
 
       await snapshot(
         replaceStackTraces(stderr.toString()).replace(
@@ -315,7 +315,7 @@ import '${fileFPath}';
           SNAPSHOT_REPLACEMENT_FILE_PATH
         ),
         new URL(
-          '../snapshots/coverage-node/script-error-stderr.ans',
+          "../snapshots/coverage-node/script-error-stderr.ans",
           import.meta.url
         )
       );
@@ -324,22 +324,22 @@ import '${fileFPath}';
     });
   });
 
-  tests.add('`coverage-node` CLI with a Node.js option, valid.', async () => {
+  tests.add("`coverage-node` CLI with a Node.js option, valid.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const filePath = join(tempDirPath, 'index.mjs');
+      const filePath = join(tempDirPath, "index.mjs");
 
       await fs.promises.writeFile(
         filePath,
-        `import { deprecate } from 'util';
+        `import { deprecate } from "util";
 
-const deprecated = deprecate(() => {}, 'Deprecated!');
+const deprecated = deprecate(() => {}, "Deprecated!");
 deprecated();
 `
       );
 
       const { stdout, stderr, status, error } = spawnSync(
-        'node',
-        [COVERAGE_NODE_CLI_PATH, '--throw-deprecation', filePath],
+        "node",
+        [COVERAGE_NODE_CLI_PATH, "--throw-deprecation", filePath],
         {
           env: {
             ...process.env,
@@ -350,24 +350,24 @@ deprecated();
 
       if (error) throw error;
 
-      strictEqual(stdout.toString(), '');
+      strictEqual(stdout.toString(), "");
       strictEqual(
-        stderr.toString().includes('DeprecationWarning: Deprecated!'),
+        stderr.toString().includes("DeprecationWarning: Deprecated!"),
         true
       );
       strictEqual(status, 1);
     });
   });
 
-  tests.add('`coverage-node` CLI with a Node.js option, invalid.', async () => {
+  tests.add("`coverage-node` CLI with a Node.js option, invalid.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const filePath = join(tempDirPath, 'index.mjs');
+      const filePath = join(tempDirPath, "index.mjs");
 
-      await fs.promises.writeFile(filePath, '1;');
+      await fs.promises.writeFile(filePath, "1;");
 
       const { stdout, stderr, status, error } = spawnSync(
-        'node',
-        [COVERAGE_NODE_CLI_PATH, '--not-a-real-option', filePath],
+        "node",
+        [COVERAGE_NODE_CLI_PATH, "--not-a-real-option", filePath],
         {
           env: {
             ...process.env,
@@ -378,12 +378,12 @@ deprecated();
 
       if (error) throw error;
 
-      strictEqual(stdout.toString(), '');
+      strictEqual(stdout.toString(), "");
 
       await snapshot(
         replaceStackTraces(stderr.toString()),
         new URL(
-          '../snapshots/coverage-node/node-option-invalid-stderr.ans',
+          "../snapshots/coverage-node/node-option-invalid-stderr.ans",
           import.meta.url
         )
       );
@@ -392,12 +392,12 @@ deprecated();
     });
   });
 
-  tests.add('`coverage-node` CLI with a missing file.', async () => {
+  tests.add("`coverage-node` CLI with a missing file.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const filePath = join(tempDirPath, 'index.mjs');
+      const filePath = join(tempDirPath, "index.mjs");
 
       const { stdout, stderr, status, error } = spawnSync(
-        'node',
+        "node",
         [COVERAGE_NODE_CLI_PATH, filePath],
         {
           env: {
@@ -409,7 +409,7 @@ deprecated();
 
       if (error) throw error;
 
-      strictEqual(stdout.toString(), '');
+      strictEqual(stdout.toString(), "");
       strictEqual(
         stderr.toString().includes(`Error: Cannot find module '${filePath}'`),
         true
@@ -418,10 +418,10 @@ deprecated();
     });
   });
 
-  tests.add('`coverage-node` CLI without arguments.', async () => {
+  tests.add("`coverage-node` CLI without arguments.", async () => {
     const { stdout, stderr, status, error } = spawnSync(
-      'node',
-      ['cli/coverage-node.mjs'],
+      "node",
+      ["cli/coverage-node.mjs"],
       {
         env: {
           ...process.env,
@@ -432,13 +432,13 @@ deprecated();
 
     if (error) throw error;
 
-    strictEqual(stdout.toString(), '');
+    strictEqual(stdout.toString(), "");
 
     await snapshot(
       replaceStackTraces(stderr.toString()),
       new URL(
         `../snapshots/coverage-node/without-arguments-coverage-${
-          coverageSupported ? 'supported' : 'unsupported'
+          coverageSupported ? "supported" : "unsupported"
         }-stderr.ans`,
         import.meta.url
       )