+++ test/private/reportCliError.test.mjs
@@ -1,43 +1,43 @@
-import { strictEqual, throws } from 'assert';
-import { spawnSync } from 'child_process';
-import fs from 'fs';
-import { join } from 'path';
-import { fileURLToPath } from 'url';
-import disposableDirectory from 'disposable-directory';
-import snapshot from 'snapshot-assertion';
-import reportCliError from '../../private/reportCliError.mjs';
-import replaceStackTraces from '../replaceStackTraces.mjs';
+import { strictEqual, throws } from "assert";
+import { spawnSync } from "child_process";
+import fs from "fs";
+import { join } from "path";
+import { fileURLToPath } from "url";
+import disposableDirectory from "disposable-directory";
+import snapshot from "snapshot-assertion";
+import reportCliError from "../../private/reportCliError.mjs";
+import replaceStackTraces from "../replaceStackTraces.mjs";
 
 const REPORT_CLI_ERROR_PATH = fileURLToPath(
-  new URL('../../private/reportCliError.mjs', import.meta.url)
+  new URL("../../private/reportCliError.mjs", import.meta.url)
 );
 
 export default (tests) => {
   tests.add(
-    '`reportCliError` with first argument `cliDescription` not a string.',
+    "`reportCliError` with first argument `cliDescription` not a string.",
     () => {
       throws(() => {
         reportCliError(true);
-      }, new TypeError('First argument `cliDescription` must be a string.'));
+      }, new TypeError("First argument `cliDescription` must be a string."));
     }
   );
 
   tests.add(
-    '`reportCliError` with a `Error` instance, with stack.',
+    "`reportCliError` with a `Error` instance, with stack.",
     async () => {
       await disposableDirectory(async (tempDirPath) => {
-        const filePath = join(tempDirPath, 'test.mjs');
+        const filePath = join(tempDirPath, "test.mjs");
 
         await fs.promises.writeFile(
           filePath,
-          `import reportCliError from '${REPORT_CLI_ERROR_PATH}';
+          `import reportCliError from "${REPORT_CLI_ERROR_PATH}";
 
-reportCliError('CLI', new Error('Message.'));
+reportCliError("CLI", new Error("Message."));
 `
         );
 
         const { stdout, stderr, status, error } = spawnSync(
-          'node',
+          "node",
           [filePath],
           {
             env: {
@@ -49,12 +49,12 @@ reportCliError('CLI', new Error('Message.'));
 
         if (error) throw error;
 
-        strictEqual(stdout.toString(), '');
+        strictEqual(stdout.toString(), "");
 
         await snapshot(
           replaceStackTraces(stderr.toString()),
           new URL(
-            '../snapshots/reportCliError/Error-instance-with-stack-stderr.ans',
+            "../snapshots/reportCliError/Error-instance-with-stack-stderr.ans",
             import.meta.url
           )
         );
@@ -65,23 +65,23 @@ reportCliError('CLI', new Error('Message.'));
   );
 
   tests.add(
-    '`reportCliError` with a `Error` instance, without stack.',
+    "`reportCliError` with a `Error` instance, without stack.",
     async () => {
       await disposableDirectory(async (tempDirPath) => {
-        const filePath = join(tempDirPath, 'test.mjs');
+        const filePath = join(tempDirPath, "test.mjs");
 
         await fs.promises.writeFile(
           filePath,
-          `import reportCliError from '${REPORT_CLI_ERROR_PATH}';
+          `import reportCliError from "${REPORT_CLI_ERROR_PATH}";
 
-const error = new Error('Message.');
+const error = new Error("Message.");
 delete error.stack;
-reportCliError('CLI', error);
+reportCliError("CLI", error);
 `
         );
 
         const { stdout, stderr, status, error } = spawnSync(
-          'node',
+          "node",
           [filePath],
           {
             env: {
@@ -93,12 +93,12 @@ reportCliError('CLI', error);
 
         if (error) throw error;
 
-        strictEqual(stdout.toString(), '');
+        strictEqual(stdout.toString(), "");
 
         await snapshot(
           replaceStackTraces(stderr.toString()),
           new URL(
-            '../snapshots/reportCliError/Error-instance-without-stack-stderr.ans',
+            "../snapshots/reportCliError/Error-instance-without-stack-stderr.ans",
             import.meta.url
           )
         );
@@ -108,23 +108,23 @@ reportCliError('CLI', error);
     }
   );
 
-  tests.add('`reportCliError` with a `CliError` instance.', async () => {
+  tests.add("`reportCliError` with a `CliError` instance.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const filePath = join(tempDirPath, 'test.mjs');
+      const filePath = join(tempDirPath, "test.mjs");
       const cliErrorPath = fileURLToPath(
-        new URL('../../private/CliError.mjs', import.meta.url)
+        new URL("../../private/CliError.mjs", import.meta.url)
       );
 
       await fs.promises.writeFile(
         filePath,
-        `import CliError from '${cliErrorPath}';
-import reportCliError from '${REPORT_CLI_ERROR_PATH}';
+        `import CliError from "${cliErrorPath}";
+import reportCliError from "${REPORT_CLI_ERROR_PATH}";
 
-reportCliError('CLI', new CliError('Message.'));
+reportCliError("CLI", new CliError("Message."));
 `
       );
 
-      const { stdout, stderr, status, error } = spawnSync('node', [filePath], {
+      const { stdout, stderr, status, error } = spawnSync("node", [filePath], {
         env: {
           ...process.env,
           FORCE_COLOR: 1,
@@ -133,12 +133,12 @@ reportCliError('CLI', new CliError('Message.'));
 
       if (error) throw error;
 
-      strictEqual(stdout.toString(), '');
+      strictEqual(stdout.toString(), "");
 
       await snapshot(
         replaceStackTraces(stderr.toString()),
         new URL(
-          '../snapshots/reportCliError/CliError-instance-stderr.ans',
+          "../snapshots/reportCliError/CliError-instance-stderr.ans",
           import.meta.url
         )
       );
@@ -147,19 +147,19 @@ reportCliError('CLI', new CliError('Message.'));
     });
   });
 
-  tests.add('`reportCliError` with a primitive value.', async () => {
+  tests.add("`reportCliError` with a primitive value.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const filePath = join(tempDirPath, 'test.mjs');
+      const filePath = join(tempDirPath, "test.mjs");
 
       await fs.promises.writeFile(
         filePath,
-        `import reportCliError from '${REPORT_CLI_ERROR_PATH}';
+        `import reportCliError from "${REPORT_CLI_ERROR_PATH}";
 
-reportCliError('CLI', '');
+reportCliError("CLI", "");
 `
       );
 
-      const { stdout, stderr, status, error } = spawnSync('node', [filePath], {
+      const { stdout, stderr, status, error } = spawnSync("node", [filePath], {
         env: {
           ...process.env,
           FORCE_COLOR: 1,
@@ -168,12 +168,12 @@ reportCliError('CLI', '');
 
       if (error) throw error;
 
-      strictEqual(stdout.toString(), '');
+      strictEqual(stdout.toString(), "");
 
       await snapshot(
         replaceStackTraces(stderr.toString()),
         new URL(
-          '../snapshots/reportCliError/primitive-value-stderr.ans',
+          "../snapshots/reportCliError/primitive-value-stderr.ans",
           import.meta.url
         )
       );