+++ test/public/analyseCoverage.test.mjs
@@ -1,30 +1,30 @@
-import { deepStrictEqual, rejects } from 'assert';
-import { spawn } from 'child_process';
-import fs from 'fs';
-import { join } from 'path';
-import disposableDirectory from 'disposable-directory';
-import childProcessPromise from '../../private/childProcessPromise.mjs';
-import analyseCoverage from '../../public/analyseCoverage.mjs';
+import { deepStrictEqual, rejects } from "assert";
+import { spawn } from "child_process";
+import fs from "fs";
+import { join } from "path";
+import disposableDirectory from "disposable-directory";
+import childProcessPromise from "../../private/childProcessPromise.mjs";
+import analyseCoverage from "../../public/analyseCoverage.mjs";
 
 export default (tests) => {
   tests.add(
-    '`reportCliError` with first argument `coverageDirPath` not a string.',
+    "`reportCliError` with first argument `coverageDirPath` not a string.",
     async () => {
       await rejects(
         analyseCoverage(true),
-        new TypeError('First argument `coverageDirPath` must be a string.')
+        new TypeError("First argument `coverageDirPath` must be a string.")
       );
     }
   );
 
   tests.add(
-    '`analyseCoverage` ignores `node_modules` directory files.',
+    "`analyseCoverage` ignores `node_modules` directory files.",
     async () => {
       await disposableDirectory(async (tempDirPath) => {
-        const coverageDirPath = join(tempDirPath, 'coverage');
-        const nodeModulesDirPath = join(tempDirPath, 'node_modules');
-        const nodeModulesModuleName = 'a';
-        const nodeModulesModuleMainFileName = 'index.mjs';
+        const coverageDirPath = join(tempDirPath, "coverage");
+        const nodeModulesDirPath = join(tempDirPath, "node_modules");
+        const nodeModulesModuleName = "a";
+        const nodeModulesModuleMainFileName = "index.mjs";
         const nodeModulesModuleMainFilePath = join(
           nodeModulesDirPath,
           nodeModulesModuleName,
@@ -35,11 +35,11 @@ export default (tests) => {
         await fs.promises.mkdir(
           join(nodeModulesDirPath, nodeModulesModuleName)
         );
-        await fs.promises.writeFile(nodeModulesModuleMainFilePath, '1;');
+        await fs.promises.writeFile(nodeModulesModuleMainFilePath, "1;");
 
         await childProcessPromise(
-          spawn('node', [nodeModulesModuleMainFilePath], {
-            stdio: 'inherit',
+          spawn("node", [nodeModulesModuleMainFilePath], {
+            stdio: "inherit",
             env: { ...process.env, NODE_V8_COVERAGE: coverageDirPath },
           })
         );
@@ -54,18 +54,18 @@ export default (tests) => {
     }
   );
 
-  tests.add('`analyseCoverage` ignores `test` directory files.', async () => {
+  tests.add("`analyseCoverage` ignores `test` directory files.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const coverageDirPath = join(tempDirPath, 'coverage');
-      const dirPath = join(tempDirPath, 'test');
-      const filePath = join(dirPath, 'index.mjs');
+      const coverageDirPath = join(tempDirPath, "coverage");
+      const dirPath = join(tempDirPath, "test");
+      const filePath = join(dirPath, "index.mjs");
 
       await fs.promises.mkdir(dirPath);
-      await fs.promises.writeFile(filePath, '1;');
+      await fs.promises.writeFile(filePath, "1;");
 
       await childProcessPromise(
-        spawn('node', [filePath], {
-          stdio: 'inherit',
+        spawn("node", [filePath], {
+          stdio: "inherit",
           env: { ...process.env, NODE_V8_COVERAGE: coverageDirPath },
         })
       );
@@ -80,17 +80,17 @@ export default (tests) => {
   });
 
   tests.add(
-    '`analyseCoverage` ignores files with `.test` prefixed before the extension.',
+    "`analyseCoverage` ignores files with `.test` prefixed before the extension.",
     async () => {
       await disposableDirectory(async (tempDirPath) => {
-        const coverageDirPath = join(tempDirPath, 'coverage');
-        const filePath = join(tempDirPath, 'index.test.mjs');
+        const coverageDirPath = join(tempDirPath, "coverage");
+        const filePath = join(tempDirPath, "index.test.mjs");
 
-        await fs.promises.writeFile(filePath, '1;');
+        await fs.promises.writeFile(filePath, "1;");
 
         await childProcessPromise(
-          spawn('node', [filePath], {
-            stdio: 'inherit',
+          spawn("node", [filePath], {
+            stdio: "inherit",
             env: { ...process.env, NODE_V8_COVERAGE: coverageDirPath },
           })
         );
@@ -105,16 +105,16 @@ export default (tests) => {
     }
   );
 
-  tests.add('`analyseCoverage` ignores files named `test`.', async () => {
+  tests.add("`analyseCoverage` ignores files named `test`.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const coverageDirPath = join(tempDirPath, 'coverage');
-      const filePath = join(tempDirPath, 'test.mjs');
+      const coverageDirPath = join(tempDirPath, "coverage");
+      const filePath = join(tempDirPath, "test.mjs");
 
-      await fs.promises.writeFile(filePath, '1;');
+      await fs.promises.writeFile(filePath, "1;");
 
       await childProcessPromise(
-        spawn('node', [filePath], {
-          stdio: 'inherit',
+        spawn("node", [filePath], {
+          stdio: "inherit",
           env: { ...process.env, NODE_V8_COVERAGE: coverageDirPath },
         })
       );
@@ -128,16 +128,16 @@ export default (tests) => {
     });
   });
 
-  tests.add('`analyseCoverage` with 1 covered file.', async () => {
+  tests.add("`analyseCoverage` with 1 covered file.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const coverageDirPath = join(tempDirPath, 'coverage');
-      const filePath = join(tempDirPath, 'index.mjs');
+      const coverageDirPath = join(tempDirPath, "coverage");
+      const filePath = join(tempDirPath, "index.mjs");
 
-      await fs.promises.writeFile(filePath, '1;');
+      await fs.promises.writeFile(filePath, "1;");
 
       await childProcessPromise(
-        spawn('node', [filePath], {
-          stdio: 'inherit',
+        spawn("node", [filePath], {
+          stdio: "inherit",
           env: { ...process.env, NODE_V8_COVERAGE: coverageDirPath },
         })
       );
@@ -151,19 +151,19 @@ export default (tests) => {
     });
   });
 
-  tests.add('`analyseCoverage` with 1 uncovered file.', async () => {
+  tests.add("`analyseCoverage` with 1 uncovered file.", async () => {
     await disposableDirectory(async (tempDirPath) => {
-      const coverageDirPath = join(tempDirPath, 'coverage');
-      const filePath = join(tempDirPath, 'index.mjs');
+      const coverageDirPath = join(tempDirPath, "coverage");
+      const filePath = join(tempDirPath, "index.mjs");
 
       await fs.promises.writeFile(
         filePath,
-        'function a() {}; function b() {};'
+        "function a() {}; function b() {};"
       );
 
       await childProcessPromise(
-        spawn('node', [filePath], {
-          stdio: 'inherit',
+        spawn("node", [filePath], {
+          stdio: "inherit",
           env: { ...process.env, NODE_V8_COVERAGE: coverageDirPath },
         })
       );